Fix issue regarding Mission Result whitespace
In Freelancer, there's the following string: Mission Result: (Success/Failed). It gets displayed after completing or failing a mission. In this string is a hidden newline, which can be seen when viewing the mission log. In addition to the newline, there's also a space between the semicolon and the final word. JFLP has removed this space because it may looked unnecessary. This however is not true as this space is actually shown in the Mission Result message after completing the mission. Because of the removal, it looks like: Mission Result:(Success/Failed). In the TSR addon, the space has been readded. However, after that there was another issue. While it looked fine in the Message success popup, in the log it showed a space after the line break. In this commit, this space has been placed before the newline character, so the space cannot be seen anymore in the log. It also still looks correct in the message popup