Undo USV transport name update
In one of my previous commits I removed 'Transport' from the USV transport names from mission 1. I did this because the factions of these transports were supposed to be Universal Shipping, but if you changed that, then the names would be too long for the contact list. Since this is an HDE-specific change, I'll undo it for the TSR repo.